[Lab 10a] Ex - Moved Store files in subfolder
