+++ app/src/main/java/com/archaeologicalfieldwork/models/mem/SpotMemStore.kt
@@ -1,5 +1,7 @@
-package com.archaeologicalfieldwork.models
+package com.archaeologicalfieldwork.models.mem
 
+import com.archaeologicalfieldwork.models.SpotModel
+import com.archaeologicalfieldwork.models.SpotStore
 import java.util.ArrayList
 import org.jetbrains.anko.info
 import org.jetbrains.anko.AnkoLogger
